+++ C++/Data-Structures/STACKS/stacks.cpp
@@ -1,72 +1,9 @@
-#include <iostream>
+#include "Stack.hpp"
 using namespace std;
-class Stack
-{
-    int tp;
-    int arr[5];
-
-    public:
-    Stack()
-    {
-        tp=-1;
-    }
-    ~Stack()
-    {
-        
-    }
-    void push(int num)
-    {
-        //check if stack is full or not
-        if(tp == 4)
-        {
-            cout << "STACK OVERFLOW" << endl;
-        }
-        else
-        {
-            tp++;
-            arr[tp] = num;
-        }
-    }
-    void pop()
-    {
-        //check if stack is empty or not
-        if(tp == -1)
-        {
-            cout << "STACK UNDERFLOW" << endl;
-        }
-        else
-        {
-            tp--;
-        }
-    }
-    void top()
-    {
-        cout << arr[tp];
-    }
-    void size()
-    {
-        cout << tp+1;
-    }
-    void print()
-    {
-        for(int i=tp;i>=0;i--)
-        {
-            cout << arr[i] << endl; 
-        }
-    }
-    bool empty()
-    {
-        if(tp==-1)
-        {
-            return true;
-        }
-        return false;
-    }
-};
 
 int main()
 {
-    Stack s;
+    Stack<int> s;
     s.push(1);
     s.push(2);
     s.push(3);
@@ -76,6 +13,14 @@ int main()
     s.pop();
     //s.size();
     //s.print();
-    cout << s.empty();
+    // cout << s.empty();
+    // Stack<string> str;
+    // str.push("hello");
+    // str.push("world");
+    // str.print();
+    Stack<int> s1;
+    s = s1;
+    s1.print();
+    //cout << s1.size();
     return 0;
 }
